implement request stub
